Refactor: Page폴더 생성, AdminPage에서 ProductManage 컴포넌트 분리
