+++ src/refactoring/hooks/useProduct.ts
@@ -3,11 +3,7 @@ import { Product } from "../../types.ts";
 
 /**
  * 엔티티(상품)을 다루는 훅
- * 
  *액션: updateProduct, addProduct
-
- *계산:
-
  *데이터: initialProducts(props), products(state)*/
 export const useProducts = (initialProducts: Product[]) => {
   const [products, setProducts] = useState<Product[]>(initialProducts);
