+++ src/refactoring/pages/CartPage.tsx
@@ -30,6 +30,7 @@ export const CartPage = ({ products, coupons }: Props) => {
     <div className="container mx-auto p-4">
       <h1 className="text-3xl font-bold mb-6">장바구니</h1>
       <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+        {/* 상품목록 컴포넌트  */}
         <div>
           <h2 className="text-2xl font-semibold mb-4">상품 목록</h2>
           <div className="space-y-2">
@@ -90,8 +91,8 @@ export const CartPage = ({ products, coupons }: Props) => {
           </div>
         </div>
         <div>
+          {/* 장바구니에 추가한 내역 컴포넌트 */}
           <h2 className="text-2xl font-semibold mb-4">장바구니 내역</h2>
-
           <div className="space-y-2">
             {cart.map((item) => {
               const appliedDiscount = getMaxApplicableDiscount(item);
@@ -141,6 +142,7 @@ export const CartPage = ({ products, coupons }: Props) => {
             })}
           </div>
 
+          {/* 쿠폰 적용 내역 */}
           <div className="mt-6 bg-white p-4 rounded shadow">
             <h2 className="text-2xl font-semibold mb-2">쿠폰 적용</h2>
             <select
@@ -168,6 +170,7 @@ export const CartPage = ({ products, coupons }: Props) => {
             )}
           </div>
 
+          {/* 주문요악 컴포넌트 */}
           <div className="mt-6 bg-white p-4 rounded shadow">
             <h2 className="text-2xl font-semibold mb-2">주문 요약</h2>
             <div className="space-y-1">
